MyProfile package reorganization
