+++ app/src/main/java/com/pepetech/squadhouse/activities/MyProfile/MyProfileActivity.java
@@ -32,10 +32,11 @@ import com.pepetech.squadhouse.activities.Explore.ExploreUserActivity;
 import com.pepetech.squadhouse.activities.FollowersActivity;
 import com.pepetech.squadhouse.activities.Following.FollowingActivity;
 import com.pepetech.squadhouse.activities.Login.LoginActivity;
-import com.pepetech.squadhouse.activities.MyProfile.helpers.BiographyBottomSheetDialog;
-import com.pepetech.squadhouse.activities.MyProfile.helpers.UpdateFullNameActivity;
-import com.pepetech.squadhouse.activities.MyProfile.helpers.UpdateProfileImageActivity;
-import com.pepetech.squadhouse.activities.MyProfile.helpers.UpdateUsernameActivity;
+import com.pepetech.squadhouse.activities.MyProfile.adapters.MyClubsAdapter;
+import com.pepetech.squadhouse.activities.MyProfile.fragments.BiographyBottomSheetDialog;
+import com.pepetech.squadhouse.activities.MyProfile.update.UpdateFullNameActivity;
+import com.pepetech.squadhouse.activities.MyProfile.update.UpdateProfileImageActivity;
+import com.pepetech.squadhouse.activities.MyProfile.update.UpdateUsernameActivity;
 import com.pepetech.squadhouse.activities.SettingsActivity;
 import com.pepetech.squadhouse.models.Club;
 import com.pepetech.squadhouse.models.Follow;
