+++ app/src/main/java/com/pepetech/squadhouse/activities/MyProfile/update/UpdateFullNameActivity.java
@@ -1,4 +1,4 @@
-package com.pepetech.squadhouse.activities.MyProfile.helpers;
+package com.pepetech.squadhouse.activities.MyProfile.update;
 
 import android.os.Bundle;
 import android.util.Log;
